+++ tests/test_kmeans.py
@@ -38,23 +38,27 @@ def initialize_kmean():
 
 
 class TestKmeans(unittest.TestCase):
-    def test_kmean_force_n_clusters(self):
-        Z = np.arange(20).reshape((5, 4))
-        row_clusters = np.array([0, 0, 1, 1, 1])
-        col_clusters = np.array([0, 1, 1, 2])
-        n_row_cluster, n_col_cluster = 1, 1
-        k_range = range(1, 3)
-        kmean_max_iter = 100
-        km = Kmeans(
-            Z=Z,
-            row_clusters=row_clusters,
-            col_clusters=col_clusters,
-            n_row_clusters=n_row_cluster,
-            n_col_clusters=n_col_cluster,
-            k_range=k_range,
-            kmean_max_iter=kmean_max_iter)
-        self.assertEqual(2, km.n_row_clusters)
-        self.assertEqual(3, km.n_col_clusters)
+    def test_kvalues_exceed_number_of_coclusters(self):
+        with self.assertRaises(ValueError):
+            Kmeans(
+                Z=np.random.random((6, 4)),
+                row_clusters=[0, 0, 1, 1, 2, 2],
+                col_clusters=[0, 0, 1, 1],
+                n_row_clusters=3,
+                n_col_clusters=2,
+                k_range=range(1, 8),
+            )
+
+    def test_kvalues_exceed_number_of_coclusters_populated(self):
+        with self.assertRaises(ValueError):
+            Kmeans(
+                Z=np.random.random((6, 4)),
+                row_clusters=[0, 0, 1, 1, 2, 2],
+                col_clusters=[0, 0, 1, 1],
+                n_row_clusters=4,
+                n_col_clusters=2,
+                k_range=range(1, 8),
+            )
 
     def test_statistic_mesures_mean(self):
         km = initialize_kmean()
@@ -72,3 +76,35 @@ class TestKmeans(unittest.TestCase):
         km = initialize_kmean()
         km.compute()
         self.assertTrue(all(np.isnan(km.results.cl_mean_centroids[2, :])))
+
+    def test_kvalue_does_not_depend_on_krange_order(self):
+        # 4 co-clusters, 2 clusters
+        Z = np.array([
+            [0, 0, 1],
+            [0, 0, 1],
+            [1, 1, 0]
+        ])
+        row_cluseters = np.array([0, 0, 1])
+        col_clusters = np.array([0, 0, 1])
+        km = Kmeans(
+            Z=Z,
+            row_clusters=row_cluseters,
+            col_clusters=col_clusters,
+            n_row_clusters=2,
+            n_col_clusters=2,
+            k_range=range(1, 5),
+            var_thres=2.0
+        )
+        res = km.compute()
+        self.assertEqual(res.k_value, 2)
+        km = Kmeans(
+            Z=Z,
+            row_clusters=row_cluseters,
+            col_clusters=col_clusters,
+            n_row_clusters=2,
+            n_col_clusters=2,
+            k_range=range(4, 0, -1),
+            var_thres=2.0
+        )
+        res = km.compute()
+        self.assertEqual(res.k_value, 2)